# Parallelize per-contract storage updates