Substantial speed ups

before 

```
omer@omer-ThinkPad-E14-Gen-3:~/Documents/juno$ go test ./... -bench=MainnetBlock -benchmem
PASS
ok      github.com/NethermindEth/juno/adapters/feeder2core      0.402s
goos: linux
goarch: amd64
pkg: github.com/NethermindEth/juno/blockchain
cpu: AMD Ryzen 7 5700U with Radeon Graphics         
BenchmarkMainnetBlock-16           10         917313646 ns/op        18574441 B/op     434711 allocs/op
```

after

```
omer@omer-ThinkPad-E14-Gen-3:~/Documents/juno$ go test ./... -bench=MainnetBlock -benchmem
PASS
ok      github.com/NethermindEth/juno/adapters/feeder2core      0.385s
goos: linux
goarch: amd64
pkg: github.com/NethermindEth/juno/blockchain
cpu: AMD Ryzen 7 5700U with Radeon Graphics         
BenchmarkMainnetBlock-16           10         544530601 ns/op        21892010 B/op     464761 allocs/op
PASS
```

we can optimize for memory later